+++ flowmachine/tests/test_labelled_flows.py
@@ -8,6 +8,9 @@ import pytest
 
 from json import loads
 
+from flowmachine.features.location.redacted_labelled_flows import (
+    RedactedLabelledFlows,
+)
 from flowmachine.features.location.labelled_flows import LabelledFlows
 from flowmachine.core import make_spatial_unit
 from flowmachine.features import SubscriberHandsetCharacteristic
@@ -238,3 +241,34 @@ def test_geojson_multi_labels(multi_labelled_flows):
         brand in dict["features"][0]["properties"]["outflows"]["Feature"].keys()
         for brand in ["Nokia", "Apple", "Sony"]
     )
+    assert (
+        dict["features"][0]["properties"]["outflows"]["Feature"]["Nokia"]["524 3 08 43"]
+        == 1
+    )
+
+
+def test_redacted_labelled_flows(labelled_flows, get_dataframe):
+    redacted_flows = RedactedLabelledFlows(
+        labelled_flows=labelled_flows, redaction_threshold=3
+    )
+    df = get_dataframe(redacted_flows)
+    assert len(df) < 10
+    assert df.value.max() > 3
+
+
+def test_redacted_multi_labelled_flows(multi_labelled_flows, get_dataframe):
+    redacted_flows = RedactedLabelledFlows(
+        labelled_flows=multi_labelled_flows, redaction_threshold=1
+    )
+    df = get_dataframe(redacted_flows)
+    assert len(df) == 19
+    assert df.value.max() > 1
+
+
+def test_redacted_spatial_unit_validation():
+    with pytest.raises(ValueError, match="spatial unit"):
+        RedactedLabelledFlows(
+            labelled_flows=SubscriberHandsetCharacteristic(
+                "2016-01-01", "2016-01-02", characteristic="hnd_type"
+            )
+        )